cleanup / refactor
